backend/requirements.txt
# Core Framework - FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML Dependencies
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0

# Gemini API - MANDATORY for awareness tips
google-generativeai==0.3.2

# Utilities
numpy>=1.24.0
python-dotenv==1.0.0
